+++ src/args.rs
@@ -3,11 +3,11 @@ use std::str::FromStr;
 use std::collections::{HashMap, HashSet};
 
 // The arguments returned by parse.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct Args {
     pub positional: Vec<String>,
     // Maps parameter names to values.
-    pub(crate) arguments: HashMap<String, String>,
+    pub(crate) arguments: HashMap<String, Option<String>>,
     pub(crate) flags: HashSet<String>,
 }
 
@@ -20,14 +20,17 @@ impl Args {
         return self.flags.contains(flag);
     }
 
-    pub fn get<T>(&self, param: &str) -> Option<T> where T: FromStr, <T as FromStr>::Err: Debug {
+    // TODO: This should be in the parser, so that we can display a help message on conversion failure.
+    pub fn get<T>(&self, param: &str) -> T where T: FromStr, <T as FromStr>::Err: Debug {
         match self.arguments.get(param) {
-            Some(value) => {
-                return Some(value.parse::<T>().expect(
+            Some(Some(value)) => {
+                return value.parse::<T>().expect(
                     &format!("Could not covert value ({}) for parameter {}", value, param)
-                ));
+                );
             },
-            None => return None,
+            None => panic!("Parameter {} unrecognized", param),
+            // Unreachable because the parser ensures that every argument has some value.
+            _ => unreachable!(),
         }
     }
 }